fixes the lab report link for lab search results
