+++ apps/modernization-ui/src/apps/search/event/components/InvestigationSearch/InvestigationResults.spec.tsx
@@ -75,10 +75,10 @@ describe('InvestigationResults component tests', () => {
         expect(getByText('Doe, John')).toBeInTheDocument();
         expect(getByText('2019 Novel Coronavirus')).toBeInTheDocument();
         expect(getByText('CAS10003001GA99')).toBeInTheDocument();
-        expect(getByText('1/1/1990')).toBeInTheDocument();
+        expect(getByText('01/01/1990')).toBeInTheDocument();
         expect(getByText(`(${age} years)`)).toBeInTheDocument();
         expect(getByText('63000')).toBeInTheDocument();
-        expect(getByText('7/21/2023')).toBeInTheDocument();
+        expect(getByText('07/21/2023')).toBeInTheDocument();
         expect(getByText('OPEN')).toBeInTheDocument();
         expect(getByText('Clayton County')).toBeInTheDocument();
         expect(getByText('2019 Novel Coronavirus').closest('a')).toHaveAttribute('href', '#');
