+++ apps/modernization-ui/src/apps/search/event/components/LabReportSearch/LabReportResults.spec.tsx
@@ -92,10 +92,10 @@ describe('LabReportResults component tests', () => {
         expect(getByText('Doe, John')).toBeInTheDocument();
         expect(getByText('Acid-Fast Stain = abnormal')).toBeInTheDocument();
         expect(getByText('OBS10001008GA01')).toBeInTheDocument();
-        expect(getByText('1/1/1990')).toBeInTheDocument();
+        expect(getByText('01/01/1990')).toBeInTheDocument();
         expect(getByText(`(${age} years)`)).toBeInTheDocument();
         expect(getByText('63000')).toBeInTheDocument();
-        expect(getByText('7/27/2023')).toBeInTheDocument();
+        expect(getByText('07/27/2023')).toBeInTheDocument();
         expect(getByText('Piedmont Hospital')).toBeInTheDocument();
         expect(getByText('Lab Report').closest('a')).toHaveAttribute('href', '#');
     });
