+++ apps/modernization-ui/src/apps/search/event/components/LabReportSearch/LabReportResults.tsx
@@ -1,13 +1,11 @@
 import { Grid, Pagination } from '@trussworks/react-uswds';
-import { useEffect, useRef } from 'react';
 import { LabReport, OrganizationParticipation, PersonParticipation } from 'generated/graphql/schema';
-import { calculateAge } from 'utils/util';
 import 'apps/search/advancedSearch/AdvancedSearch.scss';
-import { useNavigate } from 'react-router';
 import { ClassicLink } from 'classic';
 import { NoData } from 'components/NoData';
-import { formattedName } from 'utils';
 import { SearchCriteriaContext } from 'providers/SearchCriteriaContext';
+import { PatientDetails } from 'apps/search/event/components/PatientDetails';
+import { internalizeDate } from 'date';
 
 type LabReportResultsProps = {
     data: [LabReport];
@@ -17,246 +15,16 @@ type LabReportResultsProps = {
 };
 
 export const LabReportResults = ({ data, totalResults, handlePagination, currentPage }: LabReportResultsProps) => {
-    const searchItemsRef: any = useRef();
-    const navigate = useNavigate();
-
-    // Update 'width' and 'height' when the window resizes
-    useEffect(() => {
-        window.addEventListener('resize', getListSize);
-    }, []);
-
-    useEffect(() => {
-        getListSize();
-    }, [data]);
-
-    const getListSize = () => {
-        return searchItemsRef.current?.clientHeight;
-    };
-
     const handleNext = (page: number) => {
         handlePagination(page);
     };
 
-    const formatDate = (date: Date) => {
-        return new Date(date).toLocaleDateString('en-US', {
-            timeZone: 'UTC'
-        });
-    };
-
-    const getPatient = (labReport: LabReport): PersonParticipation | undefined | null => {
-        return labReport.personParticipations?.find(
-            (p) => p?.typeDescTxt === 'Patient subject' || p?.typeCd === 'PATSBJ'
-        );
-    };
-
-    const getOrderingProvidorName = (labReport: LabReport): string | undefined => {
-        const provider = labReport.personParticipations?.find((p) => p?.typeCd === 'ORD' && p?.personCd === 'PRV');
-        if (provider) {
-            return `${provider.firstName} ${provider.lastName}`;
-        } else {
-            return undefined;
-        }
-    };
-
-    const getReportingFacility = (labReport: LabReport): OrganizationParticipation | undefined | null => {
-        return labReport.organizationParticipations?.find((o) => o?.typeCd === 'AUT');
-    };
-
-    const getDescription = (labReport: LabReport): string => {
-        // TODO - there could be multiple tests associated with one lab report. How to display them in UI
-        const observation = labReport.observations?.find((o) => o?.altCd && o?.displayName && o?.cdDescTxt);
-        if (observation) {
-            return `${observation.cdDescTxt} = ${observation.displayName}`;
-        } else {
-            return 'No Data';
-        }
-    };
-
-    const buildPatientDetails = (labReport: LabReport) => {
-        const patient = getPatient(labReport);
-        let name = '';
-        let birthDate: string | undefined;
-        let age: string | undefined;
-        let sex: string | undefined;
-        if (patient) {
-            name =
-                !patient.lastName && !patient.firstName
-                    ? `No Data`
-                    : formattedName(patient?.lastName ?? '', patient?.firstName ?? '');
-            if (patient.birthTime) {
-                birthDate = formatDate(patient.birthTime);
-                age = calculateAge(new Date(patient.birthTime));
-            }
-            sex = patient.currSexCd === 'M' ? 'Male' : patient.currSexCd === 'F' ? 'Female' : 'Unknown';
-        }
-
-        const redirectPatientProfile = async () => {
-            navigate(`/patient-profile/${patient?.shortId}`);
-        };
-
-        return (
-            <Grid row gap={3}>
-                <Grid col={12} className="margin-bottom-2">
-                    <p className="margin-0 text-normal text-gray-50 search-result-item-label">LEGAL NAME</p>
-                    {name ? (
-                        <a
-                            onClick={redirectPatientProfile}
-                            className="margin-0 font-sans-md margin-top-05 text-bold text-primary word-break"
-                            style={{ wordBreak: 'break-word', cursor: 'pointer' }}>
-                            {name}
-                        </a>
-                    ) : (
-                        <NoData />
-                    )}
-                </Grid>
-                <Grid col={12} className="margin-bottom-2">
-                    <div className="grid-row flex-align-center">
-                        <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
-                            DATE OF BIRTH
-                        </p>
-                        <p className="margin-0 font-sans-1xs text-normal">
-                            <>
-                                {birthDate ? birthDate : <NoData />}
-                                <span className="font-sans-2xs"> {age ? `(${age})` : ''}</span>
-                            </>
-                        </p>
-                    </div>
-                    <div className="grid-row flex-align-center">
-                        <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">SEX</p>
-                        <p className="margin-0 font-sans-1xs text-normal">{sex ? sex : <NoData />}</p>
-                    </div>
-                    <div className="grid-row flex-align-center">
-                        <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
-                            PATIENT ID
-                        </p>
-                        <p className="margin-0 font-sans-1xs text-normal">{patient?.shortId}</p>
-                    </div>
-                </Grid>
-            </Grid>
-        );
-    };
-
     return (
         <div className="margin-x-4">
-            <div ref={searchItemsRef}>
+            <div>
                 {data &&
                     data?.length > 0 &&
-                    data?.map((item, index: number) => (
-                        <div
-                            key={index}
-                            className="padding-x-3 padding-top-3 padding-bottom-2 margin-bottom-3 bg-white border border-base-light radius-md">
-                            <Grid row gap={3}>
-                                <Grid col={4}>{buildPatientDetails(item)}</Grid>
-                                <Grid col={3}>
-                                    <Grid row gap={3}>
-                                        <Grid col={12} className="margin-bottom-2">
-                                            <p className="margin-0 text-normal search-result-item-label text-gray-50">
-                                                DOCUMENT TYPE
-                                            </p>
-                                            <ClassicLink
-                                                className="margin-0 font-sans-md margin-top-05 text-bold text-primary word-break"
-                                                url={`/nbs/api/profile/${getPatient(item)?.shortId}/report/lab/${
-                                                    item.id
-                                                }`}>
-                                                Lab Report
-                                            </ClassicLink>
-                                        </Grid>
-                                        <Grid col={12} className="margin-bottom-2">
-                                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
-                                                DATE RECEIVED
-                                            </p>
-                                            <p className="margin-0 font-sans-1xs text-normal">
-                                                {formatDate(item.addTime) || 'No Data'}
-                                            </p>
-                                        </Grid>
-                                        <Grid col={12} className="margin-bottom-2">
-                                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
-                                                DESCRIPTION
-                                            </p>
-                                            {getDescription(item) === 'No Data' ? (
-                                                <NoData />
-                                            ) : (
-                                                <p className="margin-0 font-sans-1xs text-normal">
-                                                    {getDescription(item)}
-                                                </p>
-                                            )}
-                                        </Grid>
-                                    </Grid>
-                                </Grid>
-                                <Grid col={3}>
-                                    <Grid row gap={3}>
-                                        <Grid col={12} className="margin-bottom-2">
-                                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
-                                                REPORTING FACILITY
-                                            </p>
-                                            <p className="margin-0 font-sans-1xs text-normal">
-                                                {getReportingFacility(item)?.name ?? <NoData />}
-                                            </p>
-                                        </Grid>
-                                        <Grid col={12} className="margin-bottom-2">
-                                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
-                                                ORDERING PROVIDOR
-                                            </p>
-                                            <p className="margin-0 font-sans-1xs text-normal">
-                                                {getOrderingProvidorName(item) ?? <NoData />}
-                                            </p>
-                                        </Grid>
-                                        <SearchCriteriaContext.Consumer>
-                                            {({ searchCriteria }) => (
-                                                <Grid col={12} className="margin-bottom-2">
-                                                    <p className="margin-0 text-normal search-result-item-label text-gray-50">
-                                                        JURISDICTION
-                                                    </p>
-                                                    <p className="margin-0 font-sans-1xs text-normal">
-                                                        {item.jurisdictionCd ? (
-                                                            searchCriteria.jurisdictions.find(
-                                                                (j) => j.id === item.jurisdictionCd?.toString()
-                                                            )?.codeDescTxt
-                                                        ) : (
-                                                            <NoData />
-                                                        )}
-                                                    </p>
-                                                </Grid>
-                                            )}
-                                        </SearchCriteriaContext.Consumer>
-                                    </Grid>
-                                </Grid>
-                                <Grid col={2}>
-                                    <Grid row gap={3}>
-                                        <Grid col={12} className="margin-bottom-2">
-                                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
-                                                ASSOCIATED WITH
-                                            </p>
-                                            <div className="margin-0 font-sans-1xs text-normal">
-                                                {(!item.associatedInvestigations ||
-                                                    item.associatedInvestigations.length == 0) && <NoData />}
-                                                {item.associatedInvestigations &&
-                                                    item.associatedInvestigations?.length > 0 &&
-                                                    item.associatedInvestigations?.map((i, index) => (
-                                                        <div key={index}>
-                                                            <p
-                                                                className="margin-0 text-primary text-bold"
-                                                                style={{ wordBreak: 'break-word' }}>
-                                                                {i?.localId}
-                                                            </p>
-                                                            <p className="margin-0">{i?.cdDescTxt}</p>
-                                                        </div>
-                                                    ))}
-                                            </div>
-                                        </Grid>
-                                        <Grid col={12} className="margin-bottom-2">
-                                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
-                                                LOCAL ID
-                                            </p>
-                                            <p className="margin-0 font-sans-1xs text-normal">
-                                                {item.localId || <NoData />}
-                                            </p>
-                                        </Grid>
-                                    </Grid>
-                                </Grid>
-                            </Grid>
-                        </div>
-                    ))}
+                    data?.map((item, index: number) => <LabReportSearchResult item={item} key={index} />)}
             </div>
             {Boolean(totalResults && data?.length > 0) && (
                 <Pagination
@@ -269,6 +37,150 @@ export const LabReportResults = ({ data, totalResults, handlePagination, current
                     onClickPageNumber={(_, page) => handleNext(page)}
                 />
             )}
+        </div>
+    );
+};
+
+const getPatient = (labReport: LabReport): PersonParticipation | undefined | null => {
+    return labReport.personParticipations?.find((p) => p?.typeDescTxt === 'Patient subject' || p?.typeCd === 'PATSBJ');
+};
+
+const getOrderingProvidorName = (labReport: LabReport): string | undefined => {
+    const provider = labReport.personParticipations?.find((p) => p?.typeCd === 'ORD' && p?.personCd === 'PRV');
+    if (provider) {
+        return `${provider.firstName} ${provider.lastName}`;
+    } else {
+        return undefined;
+    }
+};
+
+const getReportingFacility = (labReport: LabReport): OrganizationParticipation | undefined | null => {
+    return labReport.organizationParticipations?.find((o) => o?.typeCd === 'AUT');
+};
+
+const getDescription = (labReport: LabReport): string => {
+    // TODO - there could be multiple tests associated with one lab report. How to display them in UI
+    const observation = labReport.observations?.find((o) => o?.altCd && o?.displayName && o?.cdDescTxt);
+    if (observation) {
+        return `${observation.cdDescTxt} = ${observation.displayName}`;
+    } else {
+        return 'No Data';
+    }
+};
+
+type LabReportSearchResultProps = {
+    item: LabReport;
+};
+
+const LabReportSearchResult = ({ item }: LabReportSearchResultProps) => {
+    const patient = getPatient(item);
+    return (
+        <div className="padding-x-3 padding-top-3 padding-bottom-2 margin-bottom-3 bg-white border border-base-light radius-md">
+            <Grid row gap={3}>
+                <Grid col={4}>
+                    <PatientDetails patient={patient} />
+                </Grid>
+                <Grid col={3}>
+                    <Grid row gap={3}>
+                        <Grid col={12} className="margin-bottom-2">
+                            <p className="margin-0 text-normal search-result-item-label text-gray-50">DOCUMENT TYPE</p>
+                            <ClassicLink
+                                className="margin-0 font-sans-md margin-top-05 text-bold text-primary word-break"
+                                url={`/nbs/api/profile/${getPatient(item)?.personParentUid}/report/lab/${item.id}`}>
+                                Lab Report
+                            </ClassicLink>
+                        </Grid>
+                        <Grid col={12} className="margin-bottom-2">
+                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
+                                DATE RECEIVED
+                            </p>
+                            <p className="margin-0 font-sans-1xs text-normal">
+                                {internalizeDate(item.addTime) || 'No Data'}
+                            </p>
+                        </Grid>
+                        <Grid col={12} className="margin-bottom-2">
+                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
+                                DESCRIPTION
+                            </p>
+                            {getDescription(item) === 'No Data' ? (
+                                <NoData />
+                            ) : (
+                                <p className="margin-0 font-sans-1xs text-normal">{getDescription(item)}</p>
+                            )}
+                        </Grid>
+                    </Grid>
+                </Grid>
+                <Grid col={3}>
+                    <Grid row gap={3}>
+                        <Grid col={12} className="margin-bottom-2">
+                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
+                                REPORTING FACILITY
+                            </p>
+                            <p className="margin-0 font-sans-1xs text-normal">
+                                {getReportingFacility(item)?.name ?? <NoData />}
+                            </p>
+                        </Grid>
+                        <Grid col={12} className="margin-bottom-2">
+                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
+                                ORDERING PROVIDOR
+                            </p>
+                            <p className="margin-0 font-sans-1xs text-normal">
+                                {getOrderingProvidorName(item) ?? <NoData />}
+                            </p>
+                        </Grid>
+                        <SearchCriteriaContext.Consumer>
+                            {({ searchCriteria }) => (
+                                <Grid col={12} className="margin-bottom-2">
+                                    <p className="margin-0 text-normal search-result-item-label text-gray-50">
+                                        JURISDICTION
+                                    </p>
+                                    <p className="margin-0 font-sans-1xs text-normal">
+                                        {item.jurisdictionCd ? (
+                                            searchCriteria.jurisdictions.find(
+                                                (j) => j.id === item.jurisdictionCd?.toString()
+                                            )?.codeDescTxt
+                                        ) : (
+                                            <NoData />
+                                        )}
+                                    </p>
+                                </Grid>
+                            )}
+                        </SearchCriteriaContext.Consumer>
+                    </Grid>
+                </Grid>
+                <Grid col={2}>
+                    <Grid row gap={3}>
+                        <Grid col={12} className="margin-bottom-2">
+                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
+                                ASSOCIATED WITH
+                            </p>
+                            <div className="margin-0 font-sans-1xs text-normal">
+                                {(!item.associatedInvestigations || item.associatedInvestigations.length == 0) && (
+                                    <NoData />
+                                )}
+                                {item.associatedInvestigations &&
+                                    item.associatedInvestigations?.length > 0 &&
+                                    item.associatedInvestigations?.map((i, index) => (
+                                        <div key={index}>
+                                            <p
+                                                className="margin-0 text-primary text-bold"
+                                                style={{ wordBreak: 'break-word' }}>
+                                                {i?.localId}
+                                            </p>
+                                            <p className="margin-0">{i?.cdDescTxt}</p>
+                                        </div>
+                                    ))}
+                            </div>
+                        </Grid>
+                        <Grid col={12} className="margin-bottom-2">
+                            <p className="margin-0 text-normal search-result-item-label text-gray-50 margin-right-1">
+                                LOCAL ID
+                            </p>
+                            <p className="margin-0 font-sans-1xs text-normal">{item.localId || <NoData />}</p>
+                        </Grid>
+                    </Grid>
+                </Grid>
+            </Grid>
         </div>
     );
 };
