+++ apps/modernization-ui/src/apps/search/event/components/PatientDetails/index.ts
@@ -1,0 +1,1 @@
+export * from './PatientDetails';
